+++ CHANGELOG.md
@@ -7,7 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-- None
+### Added
+
+- Option `Minimum breadcrumb width (px)` to specify minimum width for a single breadcrumb entry ([#30](https://github.com/benji300/joplin-note-tabs/issues/30))
+
+### Changed
+
+- Breadcrumbs shall be fixed (not scrolalble) ([#29](https://github.com/benji300/joplin-note-tabs/issues/29))
 
 ## [1.1.1] - 2021-01-12